Issue #1327: Move UI tests to org.mozilla.focus.ui directory.
It's not helpful to divide these tests into many subpackages because
there aren't many tests. We should also be explicit to separate
integration tests and ui tests to show a clear separation of
erns.
Ideally, the package structure for the org.mozilla.focus package
looks like:
- helpers
- integration
- ui
