+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/PageLoadTest.java
@@ -3,16 +3,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.home;
+package org.mozilla.tv.firefox.ui;
 
 import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.ViewInteraction;
+import android.support.test.espresso.contrib.RecyclerViewActions;
 import android.support.test.espresso.matcher.ViewMatchers;
+import android.support.test.espresso.web.webdriver.Locator;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.uiautomator.UiDevice;
-import android.support.test.uiautomator.UiObject;
-import android.support.test.uiautomator.UiSelector;
 
 import org.junit.After;
 import org.junit.Rule;
@@ -27,21 +26,23 @@ import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.pressImeActionButton;
 import static android.support.test.espresso.action.ViewActions.replaceText;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.RootMatchers.withDecorView;
 import static android.support.test.espresso.matcher.ViewMatchers.hasFocus;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.isNotChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static android.support.test.espresso.web.sugar.Web.onWebView;
+import static android.support.test.espresso.web.webdriver.DriverAtoms.findElement;
 import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.core.IsNot.not;
-import static org.junit.Assert.assertFalse;
+import static org.hamcrest.core.StringContains.containsString;
 
 @RunWith(AndroidJUnit4.class)
-public class PinnedTileTests {
+public class PageLoadTest {
 
     private UiDevice mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
+    private static final Integer TILE_POSITION = 1; // Google Video Search
+    private static final String TILE_WEBSITE_ELEMENT = "hplogo"; // Google logo
+    private static final String MOZILLA_URL = "mozilla.org";
+    private static final String MOZILLA_PAGE_ELEMENT = ".primary-title";
 
     @Rule
     public ActivityTestRule<MainActivity> mActivityTestRule = new SkipOnboardingMainActivityTestRule();
@@ -52,39 +53,36 @@ public class PinnedTileTests {
     }
 
     @Test
-    public void testCustomPinnedTile() {
+    public void PageLoadTest() {
+        onView(ViewMatchers.withId(R.id.tileContainer))
+                .perform(RecyclerViewActions.actionOnItemAtPosition(TILE_POSITION, click()));
+
+        onView(ViewMatchers.withId(R.id.webview))
+                .check(matches(isDisplayed()));
+
+        onWebView()
+                .withElement(findElement(Locator.ID, TILE_WEBSITE_ELEMENT));
+
+        mDevice.pressMenu();
+
+        onView(ViewMatchers.withId(R.id.navUrlInput))
+                .check(matches(isDisplayed()))
+                .check(matches(withText(containsString("google"))));
+
         onView(allOf(withId(R.id.navUrlInput), isDisplayed(), hasFocus()))
-                .perform(replaceText("example.com"))
+                .perform(replaceText(MOZILLA_URL))
                 .perform(pressImeActionButton());
 
         onView(ViewMatchers.withId(R.id.webview))
                 .check(matches(isDisplayed()));
 
+        onWebView()
+                .withElement(findElement(Locator.CSS_SELECTOR, MOZILLA_PAGE_ELEMENT));
+
         mDevice.pressMenu();
 
-        final ViewInteraction pinButton = onView(ViewMatchers.withId(R.id.pinButton))
-                .check(matches(isNotChecked()));
-
-        pinButton.perform(click());
-
-        onView(withText(R.string.notification_pinned_site))
-                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
-                .check(matches(isDisplayed()));
-
-        UiObject newTile = mDevice.findObject(new UiSelector()
-                .resourceId("org.mozilla.tv.firefox.debug:id/tile_title")
-                .text("example")
-                .enabled(true));
-
-        newTile.waitForExists(5000);
-
-        pinButton.perform(click());
-
-        onView(withText(R.string.notification_unpinned_site))
-                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
-                .check(matches(isDisplayed()));
-
-        newTile.waitUntilGone(5000);
-        assertFalse(newTile.exists());
+        onView(ViewMatchers.withId(R.id.navUrlInput))
+                .check(matches(isDisplayed()))
+                .check(matches(withText(containsString("mozilla"))));
     }
 }
