+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/PinnedTileTests.java
@@ -3,9 +3,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.screenshots;
+package org.mozilla.tv.firefox.ui;
 
 import android.support.test.InstrumentationRegistry;
+import android.support.test.espresso.ViewInteraction;
+import android.support.test.espresso.matcher.ViewMatchers;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.uiautomator.UiDevice;
@@ -13,7 +15,6 @@ import android.support.test.uiautomator.UiObject;
 import android.support.test.uiautomator.UiSelector;
 
 import org.junit.After;
-import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,26 +22,26 @@ import org.mozilla.tv.firefox.MainActivity;
 import org.mozilla.tv.firefox.R;
 import org.mozilla.tv.firefox.SkipOnboardingMainActivityTestRule;
 
-import tools.fastlane.screengrab.Screengrab;
-import tools.fastlane.screengrab.locale.LocaleTestRule;
-
 import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.pressImeActionButton;
 import static android.support.test.espresso.action.ViewActions.replaceText;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.RootMatchers.withDecorView;
 import static android.support.test.espresso.matcher.ViewMatchers.hasFocus;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.isNotChecked;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.allOf;
-
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.core.IsNot.not;
+import static org.junit.Assert.assertFalse;
 
 @RunWith(AndroidJUnit4.class)
-public class PocketErrorTest extends ScreenshotTest {
-
+public class PinnedTileTests {
 
     private UiDevice mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
-
-    @ClassRule
-    public static final LocaleTestRule localeTestRule = new LocaleTestRule();
 
     @Rule
     public ActivityTestRule<MainActivity> mActivityTestRule = new SkipOnboardingMainActivityTestRule();
@@ -51,17 +52,39 @@ public class PocketErrorTest extends ScreenshotTest {
     }
 
     @Test
-    public void showPocketTileError() {
+    public void testCustomPinnedTile() {
         onView(allOf(withId(R.id.navUrlInput), isDisplayed(), hasFocus()))
-                .perform(replaceText("firefox:error:pocketconnection"))
+                .perform(replaceText("example.com"))
                 .perform(pressImeActionButton());
 
-        UiObject errorMsg = mDevice.findObject(new UiSelector()
-                .resourceId("org.mozilla.tv.firefox.debug:id/pocketMegaTileLoadError")
+        onView(ViewMatchers.withId(R.id.webview))
+                .check(matches(isDisplayed()));
+
+        mDevice.pressMenu();
+
+        final ViewInteraction pinButton = onView(ViewMatchers.withId(R.id.pinButton))
+                .check(matches(isNotChecked()));
+
+        pinButton.perform(click());
+
+        onView(withText(R.string.notification_pinned_site))
+                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
+                .check(matches(isDisplayed()));
+
+        UiObject newTile = mDevice.findObject(new UiSelector()
+                .resourceId("org.mozilla.tv.firefox.debug:id/tile_title")
+                .text("example")
                 .enabled(true));
 
-        errorMsg.waitForExists(5000);
+        newTile.waitForExists(5000);
 
-        Screengrab.screenshot("pocket-tile-error");
+        pinButton.perform(click());
+
+        onView(withText(R.string.notification_unpinned_site))
+                .inRoot(withDecorView(not(is(mActivityTestRule.getActivity().getWindow().getDecorView()))))
+                .check(matches(isDisplayed()));
+
+        newTile.waitUntilGone(5000);
+        assertFalse(newTile.exists());
     }
 }
