+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/screenshots/PocketErrorTest.java
@@ -1,0 +1,67 @@
+/* -*- Mode: Java; c-basic-offset: 4; tab-width: 20; indent-tabs-mode: nil; -*-
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+package org.mozilla.tv.firefox.ui.screenshots;
+
+import android.support.test.InstrumentationRegistry;
+import android.support.test.rule.ActivityTestRule;
+import android.support.test.runner.AndroidJUnit4;
+import android.support.test.uiautomator.UiDevice;
+import android.support.test.uiautomator.UiObject;
+import android.support.test.uiautomator.UiSelector;
+
+import org.junit.After;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mozilla.tv.firefox.MainActivity;
+import org.mozilla.tv.firefox.R;
+import org.mozilla.tv.firefox.SkipOnboardingMainActivityTestRule;
+
+import tools.fastlane.screengrab.Screengrab;
+import tools.fastlane.screengrab.locale.LocaleTestRule;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.pressImeActionButton;
+import static android.support.test.espresso.action.ViewActions.replaceText;
+import static android.support.test.espresso.matcher.ViewMatchers.hasFocus;
+import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static org.hamcrest.Matchers.allOf;
+
+
+@RunWith(AndroidJUnit4.class)
+public class PocketErrorTest extends ScreenshotTest {
+
+
+    private UiDevice mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation());
+
+    @ClassRule
+    public static final LocaleTestRule localeTestRule = new LocaleTestRule();
+
+    @Rule
+    public ActivityTestRule<MainActivity> mActivityTestRule = new SkipOnboardingMainActivityTestRule();
+
+    @After
+    public void tearDown() {
+        mActivityTestRule.getActivity().finishAndRemoveTask();
+    }
+
+    @Test
+    public void showPocketTileError() {
+        onView(allOf(withId(R.id.navUrlInput), isDisplayed(), hasFocus()))
+                .perform(replaceText("firefox:error:pocketconnection"))
+                .perform(pressImeActionButton());
+
+        UiObject errorMsg = mDevice.findObject(new UiSelector()
+                .resourceId("org.mozilla.tv.firefox.debug:id/pocketMegaTileLoadError")
+                .enabled(true));
+
+        errorMsg.waitForExists(5000);
+
+        Screengrab.screenshot("pocket-tile-error");
+    }
+}
