+++ app/src/androidTest/java/org/mozilla/tv/firefox/ui/screenshots/TVScreenshots.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.screenshots;
+package org.mozilla.tv.firefox.ui.screenshots;
 
 import android.content.Context;
 import android.content.Intent;
